+++ js/collapsibleTabs.js
@@ -9,7 +9,7 @@
 			return this;
 		}
 		// Merge options into the defaults
-		var $settings = $.extend( {}, $.collapsibleTabs.defaults, options );
+		var settings = $.extend( {}, $.collapsibleTabs.defaults, options );
 
 		this.each( function () {
 			var $el = $( this );
@@ -17,52 +17,25 @@
 			$.collapsibleTabs.instances = ( $.collapsibleTabs.instances.length === 0 ?
 				$el : $.collapsibleTabs.instances.add( $el ) );
 			// attach the settings to the elements
-			$el.data( 'collapsibleTabsSettings', $settings );
+			$el.data( 'collapsibleTabsSettings', settings );
 			// attach data to our collapsible elements
-			$el.children( $settings.collapsible ).each( function () {
+			$el.children( settings.collapsible ).each( function () {
 				$.collapsibleTabs.addData( $( this ) );
 			} );
 		} );
 
-		// if we haven't already bound our resize hanlder, bind it now
+		// if we haven't already bound our resize handler, bind it now
 		if ( !$.collapsibleTabs.boundEvent ) {
-			$( window )
-				.delayedBind( 500, 'resize', function () {
-					$.collapsibleTabs.handleResize();
-				} );
+			$( window ).on( 'resize', $.debounce( 500, function () {
+				$.collapsibleTabs.handleResize();
+			} ) );
+			$.collapsibleTabs.boundEvent = true;
 		}
+
 		// call our resize handler to setup the page
 		$.collapsibleTabs.handleResize();
 		return this;
 	};
-	/**
-	 * Returns the amount of horizontal distance between the two tabs groups
-	 * (#left-navigation and #right-navigation), in pixels. If negative, this
-	 * means that the tabs overlap, and the value is the width of overlapping
-	 * parts.
-	 *
-	 * Used in default expandCondition and collapseCondition.
-	 *
-	 * @return {Numeric} distance/overlap in pixels
-	 */
-	function calculateTabDistance() {
-		var $leftTab, $rightTab, leftEnd, rightStart;
-
-		// In RTL, #right-navigation is actually on the left and vice versa.
-		// Hooray for descriptive naming.
-		if ( !rtl ) {
-			$leftTab = $( '#left-navigation' );
-			$rightTab = $( '#right-navigation' );
-		} else {
-			$leftTab = $( '#right-navigation' );
-			$rightTab = $( '#left-navigation' );
-		}
-
-		leftEnd = $leftTab.offset().left + $leftTab.width();
-		rightStart = $rightTab.offset().left;
-
-		return rightStart - leftEnd;
-	}
 	$.collapsibleTabs = {
 		instances: [],
 		boundEvent: null,
@@ -72,36 +45,34 @@
 			collapsible: 'li.collapsible',
 			shifting: false,
 			expandCondition: function ( eleWidth ) {
-				// If there's at least eleWidth pixels free space, expand.
-				return calculateTabDistance() >= eleWidth;
+				// If there are at least eleWidth + 1 pixels of free space, expand.
+				// We add 1 because .width() will truncate fractional values but .offset() will not.
+				return $.collapsibleTabs.calculateTabDistance() >= eleWidth + 1;
 			},
 			collapseCondition: function () {
 				// If there's an overlap, collapse.
-				return calculateTabDistance() < 0;
+				return $.collapsibleTabs.calculateTabDistance() < 0;
 			}
 		},
 		addData: function ( $collapsible ) {
-			var $settings = $collapsible.parent().data( 'collapsibleTabsSettings' );
-			if ( $settings !== null ) {
+			var settings = $collapsible.parent().data( 'collapsibleTabsSettings' );
+			if ( settings ) {
 				$collapsible.data( 'collapsibleTabsSettings', {
-					expandedContainer: $settings.expandedContainer,
-					collapsedContainer: $settings.collapsedContainer,
+					expandedContainer: settings.expandedContainer,
+					collapsedContainer: settings.collapsedContainer,
 					expandedWidth: $collapsible.width(),
 					prevElement: $collapsible.prev()
 				} );
 			}
 		},
 		getSettings: function ( $collapsible ) {
-			var $settings = $collapsible.data( 'collapsibleTabsSettings' );
-			if ( $settings === undefined ) {
+			var settings = $collapsible.data( 'collapsibleTabsSettings' );
+			if ( !settings ) {
 				$.collapsibleTabs.addData( $collapsible );
-				$settings = $collapsible.data( 'collapsibleTabsSettings' );
+				settings = $collapsible.data( 'collapsibleTabsSettings' );
 			}
-			return $settings;
+			return settings;
 		},
-		/**
-		 * @param {jQuery.Event} e
-		 */
 		handleResize: function () {
 			$.collapsibleTabs.instances.each( function () {
 				var $el = $( this ),
@@ -132,21 +103,21 @@
 			} );
 		},
 		moveToCollapsed: function ( ele ) {
-			var data, expContainerSettings, target,
+			var outerData, expContainerSettings, target,
 				$moving = $( ele );
 
-			data = $.collapsibleTabs.getSettings( $moving );
-			if ( !data ) {
+			outerData = $.collapsibleTabs.getSettings( $moving );
+			if ( !outerData ) {
 				return;
 			}
-			expContainerSettings = $.collapsibleTabs.getSettings( $( data.expandedContainer ) );
+			expContainerSettings = $.collapsibleTabs.getSettings( $( outerData.expandedContainer ) );
 			if ( !expContainerSettings ) {
 				return;
 			}
 			expContainerSettings.shifting = true;
 
 			// Remove the element from where it's at and put it in the dropdown menu
-			target = data.collapsedContainer;
+			target = outerData.collapsedContainer;
 			$moving.css( 'position', 'relative' )
 				.css( ( rtl ? 'left' : 'right' ), 0 )
 				.animate( { width: '1px' }, 'normal', function () {
@@ -154,9 +125,7 @@
 					$( this ).hide();
 					// add the placeholder
 					$( '<span class="placeholder" style="display: none;"></span>' ).insertAfter( this );
-					// XXX: 'data' is undefined here, should the 'data' from the outer scope have
-					// a different name?
-					$( this ).detach().prependTo( target ).data( 'collapsibleTabsSettings', data );
+					$( this ).detach().prependTo( target ).data( 'collapsibleTabsSettings', outerData );
 					$( this ).attr( 'style', 'display: list-item;' );
 					data = $.collapsibleTabs.getSettings( $( ele ) );
 					if ( data ) {
@@ -204,6 +173,34 @@
 					}
 				} )
 			);
+		},
+		/**
+		 * Returns the amount of horizontal distance between the two tabs groups
+		 * (#left-navigation and #right-navigation), in pixels. If negative, this
+		 * means that the tabs overlap, and the value is the width of overlapping
+		 * parts.
+		 *
+		 * Used in default expandCondition and collapseCondition.
+		 *
+		 * @return {Numeric} distance/overlap in pixels
+		 */
+		calculateTabDistance: function () {
+			var $leftTab, $rightTab, leftEnd, rightStart;
+
+			// In RTL, #right-navigation is actually on the left and vice versa.
+			// Hooray for descriptive naming.
+			if ( !rtl ) {
+				$leftTab = $( '#left-navigation' );
+				$rightTab = $( '#right-navigation' );
+			} else {
+				$leftTab = $( '#right-navigation' );
+				$rightTab = $( '#left-navigation' );
+			}
+
+			leftEnd = $leftTab.offset().left + $leftTab.width();
+			rightStart = $rightTab.offset().left;
+
+			return rightStart - leftEnd;
 		}
 	};